requirements.txt
asgiref==3.7.2
cloudinary==1.40.0
dj-database-url==2.1.0
django==5.0.6
django-allauth==0.51.0
django-cloudinary-storage==0.3.0
django-compressor==4.3
django-cors-headers==4.3.1
django-csp==3.6
django-filter==24.2
django-rest-auth==0.9.5
django-storages==1.13.2
django-request==1.6.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dj-rest-auth==2.2.7  
gunicorn==22.0.0
openai==1.35.10
pillow==10.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
setuptools==70.3.0
whitenoise==6.7.0
